Markov generation is now done in a thread
Due to high time spent there that was blocking the main thread.
