+++ package.json
@@ -21,6 +21,7 @@
     "markov-generator": "^1.2.3",
     "mongoose": "^5.11.15",
     "mongoose-field-encryption": "^4.0.1",
-    "node-telegram-bot-api": "^0.51.0"
+    "node-telegram-bot-api": "^0.51.0",
+    "workerpool": "^6.2.0"
   }
 }
